seperate js
